[Chore] Automation Interfaces File Divided into Stepper and ExplanationChip
