+++ src/components/SettingModal/TermControlButton.tsx
@@ -1,4 +1,4 @@
-import React, {useImperativeHandle} from 'react';
+import React from 'react';
 import IconButton from "@material-ui/core/IconButton";
 import ArrowDropUpIcon from "@material-ui/icons/ArrowDropUp";
 import ArrowDropDownIcon from "@material-ui/icons/ArrowDropDown";
@@ -10,26 +10,17 @@ import {StorageKeys} from "../../constants";
 import {getReduxData} from "@funcUtils/getReduxData";
 import {ReducerAutomationDto} from "@redux/modules/ControlAutomation";
 import {AvailableMachines} from "@interfaces/main";
-import {TaskNextButtonRef} from "@components/SettingModal/CustomStepper";
 
 interface TermControlButtonProp {
   machine: AvailableMachines;
 }
 
-const TermControlButton = React.forwardRef((
-    {machine}: TermControlButtonProp, ref?: React.Ref<TaskNextButtonRef>
+const TermControlButton = (
+    {machine}: TermControlButtonProp
   ) => {
   const dispatch = useDispatch();
   const singleAutomation: ReducerAutomationDto = getReduxData( StorageKeys.AUTO )[ machine ]
   const [term, setTerm] = React.useState<number>( singleAutomation.term );
-
-  useImperativeHandle(ref, () => ({
-    handleNextStep () {
-      dispatch(controlAutomation(update(singleAutomation, {
-        term: { $set: term },
-      })))
-    }
-  }))
 
   const handleTermUp = () => {
     const singleAutomation: ReducerAutomationDto = getReduxData( StorageKeys.AUTO )[ machine ]
@@ -68,6 +59,5 @@ const TermControlButton = React.forwardRef((
     </>
   )
 }
-)
 
 export default React.memo(TermControlButton);
