+++ src/interfaces/ExplanationChip.tsx
@@ -1,77 +1,10 @@
-import {AvailableAutomationType, AvailableMachines, AvailableMachineSection} from "./main";
-import {TaskNextButtonRef} from "@components/SettingModal/CustomStepper";
-import React, {ReactNode} from "react";
-import {RangeSlider} from "@components/SettingModal/RangeSlider";
-import TimeSpanPickerWrapper from "@components/SettingModal/TimePickerWrapper";
-import SettingExplanation from "@components/SettingModal/SettingExplanation";
 import {Units} from "@values/units";
 import {Translations} from "@values/translations";
 import Chip from "@material-ui/core/Chip";
 import LoopIcon from "@material-ui/icons/Loop";
+import React from "react";
 
-export interface ResponseAutomationRead {
-  lastAutomations: Automation[];
-}
-
-export interface Automation {
-  start: string[];
-  end: string[];
-  term: number;
-  enable: boolean;
-  machine: AvailableMachines;
-  automationType: AvailableAutomationType;
-  machineSection: AvailableMachineSection;
-}
-
-
-class Stepper {
-  position: string; ref: React.Ref<TaskNextButtonRef>;
-  constructor (
-    position: string | AvailableMachines,
-    ref?: React.Ref<TaskNextButtonRef>
-  ) {
-    this.position = position;
-    this.ref = ref as React.Ref<TaskNextButtonRef>;
-  }
-
-  render() {}
-}
-
-export class RangeStepper extends Stepper {
-  render(): ReactNode {
-    return <RangeSlider key={this.position}
-                        position={this.position as AvailableMachines}
-                        ref={this.ref} />
-  }
-}
-
-export class TimePickerStepper extends Stepper {
-  outerSize: number;
-  constructor(
-    position: string,
-    outerSize: number,
-    ref: React.Ref<TaskNextButtonRef>,
-  ) {
-    super(position, ref);
-    this.outerSize = outerSize;
-  }
-
-  render(): ReactNode {
-    return <TimeSpanPickerWrapper outerSize = {this.outerSize}
-                                  key={this.position}
-                                  position={this.position as AvailableMachines}
-                                  ref={this.ref} />
-  }
-}
-
-export class AutomationExplanationStepper extends Stepper {
-  render(): ReactNode {
-    return <SettingExplanation key={this.position}
-                               position={this.position} />
-  }
-}
-
-export class ExplanationChip {
+class ExplanationChip {
   valuetext<T extends number, U extends string>(value: T, status: U) { }
   onText () {  }
   offText () {  }
@@ -141,7 +74,7 @@ export class CycleExplanationChip extends ExplanationChip {
             variant="outlined"
             size="small"
             label={`${this.term} 일`} />
-      )
+    )
     this.start.forEach((s, i) => {
       result.push (
         <Chip key={`${s} - ${this.end[i]}`}
